+++ server/src/llm/prompts.ts
@@ -1,52 +1,18 @@
-// Analysis prompts and response parsers for LLM session analysis.
-// Ported from web repo (src/lib/llm/prompts.ts) with SQLite-aware message formatting.
+// Prompt template strings and generator functions for LLM session analysis.
+// Types → prompt-types.ts, constants → prompt-constants.ts,
+// formatting → message-format.ts, parsers → response-parsers.ts.
 
-export type {
-  SQLiteMessageRow,
-  SessionMetadata,
-  AnalysisResponse,
-  ParseError,
-  ParseResult,
-  PromptQualityFinding,
-  PromptQualityTakeaway,
-  PromptQualityDimensionScores,
-  PromptQualityResponse,
-} from './prompt-types.js';
 import type { SessionMetadata } from './prompt-types.js';
-
-export {
-  FRICTION_CLASSIFICATION_GUIDANCE,
-  CANONICAL_FRICTION_CATEGORIES,
-  CANONICAL_PATTERN_CATEGORIES,
-  CANONICAL_PQ_DEFICIT_CATEGORIES,
-  CANONICAL_PQ_STRENGTH_CATEGORIES,
-  CANONICAL_PQ_CATEGORIES,
-  PROMPT_QUALITY_CLASSIFICATION_GUIDANCE,
-  EFFECTIVE_PATTERN_CLASSIFICATION_GUIDANCE,
-} from './prompt-constants.js';
 import {
   FRICTION_CLASSIFICATION_GUIDANCE,
   CANONICAL_FRICTION_CATEGORIES,
   CANONICAL_PATTERN_CATEGORIES,
   CANONICAL_PQ_DEFICIT_CATEGORIES,
   CANONICAL_PQ_STRENGTH_CATEGORIES,
-  CANONICAL_PQ_CATEGORIES,
   PROMPT_QUALITY_CLASSIFICATION_GUIDANCE,
   EFFECTIVE_PATTERN_CLASSIFICATION_GUIDANCE,
 } from './prompt-constants.js';
-
-export {
-  classifyStoredUserMessage,
-  formatMessagesForAnalysis,
-  formatSessionMetaLine,
-} from './message-format.js';
 import { formatSessionMetaLine } from './message-format.js';
-
-export {
-  extractJsonPayload,
-  parseAnalysisResponse,
-  parsePromptQualityResponse,
-} from './response-parsers.js';
 
 /**
  * System prompt for session analysis.
